a Few changes to the views have been made
