+++ Kamelna/Kamelna/Views/Home Screen/HomeView.swift
@@ -216,70 +216,6 @@ struct HomeView: View {
         }
     }
 }
-
-struct StatView: View {
-    let icon: String
-    let value: String
-    let color: Color
-    
-    var body: some View {
-        VStack {
-            Image(systemName: icon)
-                .foregroundColor(color)
-            Text(value)
-                .font(.caption)
-        }
-    }
-}
-
-struct SessionButton: View {
-    let title: String
-    let icon: String
-    
-    var body: some View {
-        VStack {
-            Image(systemName: icon)
-                .font(.title2)
-                .foregroundStyle(ButtonBackGroundColor.backgroundGradient)
-            Text(title)
-                .font(.caption2)
-                .multilineTextAlignment(.center)
-                .foregroundStyle(ButtonBackGroundColor.backgroundGradient)
-        }
-        .frame(width: 70)
-    }
-}
-
-struct TabBarButton: View {
-    let title: String
-    let icon: String
-    var isActive: Bool = false
-    var badge: Int? = nil
-    
-    var body: some View {
-        VStack {
-            ZStack {
-                Image(systemName: icon)
-                    .font(.title3)
-                    .foregroundStyle(isActive ? BackgroundGradient.backgroundGradient : ButtonForeGroundColor.backgroundGradient)
-                if let badge, badge > 0 {
-                    Text("\(badge)")
-                        .font(.caption2)
-                        .foregroundColor(.white)
-                        .padding(4)
-                        .background(Color.red)
-                        .clipShape(Circle())
-                        .offset(x: 10, y: -10)
-                }
-            }
-            Text(title)
-                .font(.caption2)
-                .foregroundStyle(isActive ? BackgroundGradient.backgroundGradient : ButtonForeGroundColor.backgroundGradient)
-        }
-        .frame(maxWidth: .infinity)
-    }
-}
-
 #Preview {
     HomeView()
 }
