+++ Kamelna/Kamelna/Views/RoomChatView.swift
@@ -42,9 +42,10 @@ struct RoomChatView: View {
 
                                         Text(message.text)
                                             .padding(10)
-                                            .background(Color.blue)
+                                            .background(ButtonBackGroundColor.backgroundGradient)
                                             .cornerRadius(10)
-                                            .foregroundColor(.white)
+                                            .foregroundStyle(ButtonForeGroundColor.backgroundGradient)
+
 
                                         Text(message.timestamp.formatted(date: .omitted, time: .shortened))
                                             .font(.caption2)
@@ -59,9 +60,9 @@ struct RoomChatView: View {
 
                                         Text(message.text)
                                             .padding(10)
-                                            .background(Color.gray.opacity(0.2))
+                                            .background(UnSelectedButtonBackGroundColor.backgroundGradient)
                                             .cornerRadius(10)
-                                            .foregroundColor(.black)
+                                            .foregroundStyle(ButtonForeGroundColor.backgroundGradient)
 
                                         Text(message.timestamp.formatted(date: .omitted, time: .shortened))
                                             .font(.caption2)
